refactor: pages/ 디렉토리 삭제
