+++ apps/client/src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
-import Home from '@/pages/Home';
+import Home from '@/components/Home.tsx';
 
 export const Route = createFileRoute('/')({
   beforeLoad: ({ context }) => {
