+++ apps/client/src/routes/signup.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
-import Signup from '@/pages/Signup';
+import Signup from '@/components/Signup.tsx';
 
 export const Route = createFileRoute('/signup')({
   beforeLoad: ({ context }) => {
