--- apps/client/src/utils/authStorage.ts
@@ -1,11 +1,0 @@
-export const getStoredUser = () => {
-  return localStorage.getItem('user');
-};
-
-export const setStoredUser = (user: string | null) => {
-  if (user) {
-    localStorage.setItem('user', user);
-  } else {
-    localStorage.removeItem('user');
-  }
-};
